requirements.txt.txt
Flask==2.1.0
nltk==3.6.5
google-generativeai==0.3.1
graphviz==0.20.1
gunicorn==20.1.0